requirements_for_analyse_cmd_py.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorlog==6.8.2
cryptography==43.0.0
distro==1.9.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==2.1.0
openai==0.27.0
packaging==24.1
pdfminer.six==20240706
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyPDF2==3.0.1
python-pptx==1.0.2
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.2
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
urllib3==2.2.2
XlsxWriter==3.2.0
yarl==1.9.4
